+++ CHANGES.md
@@ -1,6 +1,7 @@
 v0.2.1 (dev)
 ============
 + fix hover in html
++ add --unknown flag for `somalier relate` to set unknown genotypes to hom-ref (useful when merging single-sample VCFs).
 
 v0.2.0
 ======